Build in point of order restriction on closed (#2553) (#2559)
* Build in point of order restriction on closed (#2553)

* Build in point of order restriction on closed

* Change setting name

* Disable point of order button if point of order restriction is turned on

---------

+++ client/src/app/site/pages/meetings/modules/list-of-speakers-content/components/list-of-speakers-content/list-of-speakers-content.component.ts
@@ -123,6 +123,8 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
 
     public pointOfOrderCategoriesEnabled: boolean = false;
 
+    public restrictPointOfOrderActions: boolean = false;
+
     @Output()
     private isListOfSpeakersEmptyEvent = new EventEmitter<boolean>();
 
@@ -159,7 +161,10 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
         this.subscriptions.push(
             this.meetingSettingsService
                 .get(`list_of_speakers_enable_point_of_order_categories`)
-                .subscribe(enabled => (this.pointOfOrderCategoriesEnabled = enabled))
+                .subscribe(enabled => (this.pointOfOrderCategoriesEnabled = enabled)),
+            this.meetingSettingsService
+                .get(`list_of_speakers_closing_disables_point_of_order`)
+                .subscribe(enabled => (this.restrictPointOfOrderActions = enabled))
         );
     }
 
@@ -255,7 +260,10 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
     public async removePointOfOrder(): Promise<void> {
         const speakerToDelete = this.findOperatorSpeaker(true);
         if (speakerToDelete) {
-            await this.speakerRepo.delete(speakerToDelete.id);
+            const title = this.translate.instant(`Are you sure you want to irrevocably remove your point of order?`);
+            if (!(this.restrictPointOfOrderActions && this.closed) || (await this.promptService.open(title))) {
+                await this.speakerRepo.delete(speakerToDelete.id);
+            }
         }
     }
